pyproject.toml
[tool.poetry]
name = "arcan"
version = "0.1.1"
description = "An AI web3 tooling platform for the decentralized customization and enhancement of AI agents"
authors = ["Carlos D. Escobar-Valbuena <carlosdavidescobar@gmail.com>"]
license = "Apache 2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
urllib3 = ">=1.26.16,<3.0.0"
fastapi = "^0.110.0"
openai = "^1.14.0"
langchain = "^0.1.16"
langchain-openai = "^0.0.8"
# uvicorn = "^0.28.0"
pydantic = ">=1.10.13,<=2.8.0"
python-dotenv = "^1.0.1"
typing-extensions = "^4.9.0"
pandas = "^2.2.1"
sqlalchemy = "^2.0.27"
psycopg2-binary = "^2.9.9"
bs4 = "^0.0.2"
selenium = "^4.18.1"
html2text = "^2024.2.26"
chromadb = "^0.4.24"
pymupdf = "^1.23.25"
wikipedia = "^1.4.0"
sendgrid = "^6.11.0"
twilio = "^9.0.0"
python-decouple = "^3.8"
python-multipart = "^0.0.9"
pyngrok = "^7.1.3"
numexpr = "^2.9.0"
langchainhub = "^0.1.15"
langchain-experimental = "^0.0.54"
langchain-community = "^0.0.32"
# langchain-together = "^0.0.2.post1"
# langchain-fireworks = "^0.1.1"
modal = "^0.61.54"
typer = "^0.9.0"
langserve = {extras = ["all"], version = ">=0.0.30"} #"^0.1.1"
langchain-groq = "^0.1.3"
firecrawl-py = "^0.0.8"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
supabase = "^2.4.5"
chainlit = "^1.1.101"
llama-index = "^0.10.37"
bcrypt = "^4.1.3"
loguru = "^0.7.2"
pydantic-settings = "^2.3.1"
asyncpg = "^0.29.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
ipykernel = "^6.23.3"
pytest = "^8.2.0"
httpx = "^0.27.0"
pytest-asyncio = "^0.23.6"

[tool.poetry.scripts]
arcan = 'arcan:cli'